refactor: consolidate utility function imports into a single index file
