+++ src/components/ChartCard/ChartCard.tsx
@@ -8,8 +8,7 @@ import { Cross1Icon } from '@radix-ui/react-icons'
 import useResponsive from '../../hooks/useResponsive'
 import { useFullscreen } from '../../hooks/useFullscreen'
 import { useMapVisualization, useMapData } from '../../hooks/useGlobalContext'
-import { getNameByCountryCode } from '../../library/utils/getNameByCountryCode'
-import { capitalize } from '../../library/utils/capitalize'
+import { getNameByCountryCode, capitalize } from '../../library/utils'
 import { NONE_VALUE, RATES_OF_CHANGE_YEARS } from '../../library/constants'
 
 import BarChartIcon from '../../assets/bar-chart.svg'
