+++ src/components/SearchBar/SearchBar.tsx
@@ -4,7 +4,7 @@ import styles from './SearchBar.module.scss'
 import DEPLogo from '../../assets/DEP-logo.jpg'
 import { NONE_VALUE, SEARCHBAR_INITIAL_VALUE } from '../../library/constants'
 import { useMapVisualization, useMapData } from '../../hooks/useGlobalContext'
-import { getNameByCountryCode } from '../../library/utils/getNameByCountryCode'
+import { getNameByCountryCode } from '../../library/utils'
 import type { CountryGeoJSONFeature } from '../../library/types'
 
 export const SearchBar = () => {
