+++ src/contexts/MapDataContext.tsx
@@ -1,9 +1,8 @@
 import React, { createContext, useState, useEffect } from 'react'
 import { useSearchParams } from 'react-router-dom'
-import { normalize } from '../library/utils/normalize'
 import { PACIFIC_COUNTRIES_NAMES } from '../library/constants'
 import type { ContiguousHotspotProperties, CountryGeoJSONFeature } from '../library/types'
-import { getNameByCountryCode } from '../library/utils/getNameByCountryCode'
+import { normalize, getNameByCountryCode } from '../library/utils'
 
 interface MapDataContextType {
   countryApiData: CountryGeoJSONFeature[] | []
